feat: manually updated reason and points are editable
refactor: ExperiencePointsTable removing duplicate
