+++ client/app/api/course/ExperiencePointsRecord.ts
@@ -33,9 +33,8 @@ export default class ExperiencePointsRecordAPI extends BaseCourseAPI {
     pageNum: number = 1,
   ): Promise<AxiosResponse<ExperiencePointsRecords>> {
     return this.client.get(
-      `${
-        this.#urlPrefix
-      }/users/${userId}/experience_points_records?filter[page_num]=${pageNum}`,
+      `${this.#urlPrefix}/users/${userId}/experience_points_records`,
+      { params: { 'filter[page_num]': pageNum } },
     );
   }
 
@@ -51,9 +50,10 @@ export default class ExperiencePointsRecordAPI extends BaseCourseAPI {
   update(
     params: UpdateExperiencePointsRecordPatchData,
     recordId: number,
+    studentId?: number,
   ): Promise<AxiosResponse<ExperiencePointsRecordListData>> {
     const url = `${this.#urlPrefix}/users/${
-      this.courseUserId
+      this.courseUserId ?? studentId
     }/experience_points_records/${recordId}`;
     return this.client.patch(url, params);
   }
@@ -61,9 +61,9 @@ export default class ExperiencePointsRecordAPI extends BaseCourseAPI {
   /**
    * Delete an experience points record for a user
    */
-  delete(recordId: number): Promise<AxiosResponse<void>> {
+  delete(recordId: number, studentId?: number): Promise<AxiosResponse<void>> {
     const url = `${this.#urlPrefix}/users/${
-      this.courseUserId
+      this.courseUserId ?? studentId
     }/experience_points_records/${recordId}`;
     return this.client.delete(url);
   }
